+++ lib/toc.ts
@@ -1,0 +1,9 @@
+export interface TocItem {
+  level: number;
+  text: string;
+  href: string;
+}
+
+export interface MobTocProps {
+  tocs: TocItem[];
+}